qt gui: use double instead of float in automatic data type translation
When translating a SC slot into a QVariant, always translate a Float
into a double, since there is no way to differentiate between a float
and a double in SC.

Accordingly, all the invokable methods of widgets have to change
signature to use double instead of float, or the automatic signature
matching will fail.

+++ QtCollider/widgets/QcNumberBox.h
@@ -87,8 +87,8 @@ class QcNumberBox : public QLineEdit, QcHelper, QcAbstractStepValue
     void setMaxDecimals( int );
 
   public Q_SLOTS:
-    void increment( float factor );
-    void decrement( float factor );
+    void increment( double factor );
+    void decrement( double factor );
 
   Q_SIGNALS:
     void scrolled( int steps );
